// src/app/models/vo/PregAsigResp.ts
export class PregAsigResp {
    id_pregunta: number;
    id_respuestaasignada: number;
    minuto: number;
    segundo: number;
    tiempo: string;


    constructor(
        id_pregunta?: number,
        id_respuestaasignada?: number,
        minuto?: number,
        segundo?: number,
        tiempo?: string




    ) {}
}